Move additions to medical.prescription.order.line model into medical_prescription module
