+++ website_portal_medical_prescription_order_line/__init__.py
@@ -3,4 +3,3 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import controllers
-from . import models
