--- website_portal_medical_prescription_order_line/models/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2016 LasLabs Inc.
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import medical_prescription_order_line
